requirements.txt
opencv-python
mediapipe
pygame
PyOpenGL
numpy
